% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plots.R
\name{continuousBaroPressure}
\alias{continuousBaroPressure}
\title{Generate barometric pressure plot}
\usage{
continuousBaroPressure(site)
}
\arguments{
\item{site}{Three- or four-digit site code (e.g., KMMH, 315)}
}
\value{
Ggplot object
}
\description{
Generate barometric pressure plot
}
