% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plots.R
\name{continuousRawLevel}
\alias{continuousRawLevel}
\title{Generate well raw water level plot (not baro compensated)}
\usage{
continuousRawLevel(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Ggplot object
}
\description{
Generate well raw water level plot (not baro compensated)
}
