% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plots.R
\name{continuousWaterLevel}
\alias{continuousWaterLevel}
\title{Generate well water level plot}
\usage{
continuousWaterLevel(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Ggplot object
}
\description{
Generate well water level plot
}
