% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plots.R
\name{continuousWaterTemperature}
\alias{continuousWaterTemperature}
\title{Generate well water temperature plot}
\usage{
continuousWaterTemperature(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Ggplot object
}
\description{
Generate well water temperature plot
}
