% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tables.R
\name{daysWithinSurface}
\alias{daysWithinSurface}
\title{Generate table showing the count and percentage of days per water year where water is within 10 cm and 50 cm of the ground surface (for years where there are at least 300 days of data)}
\usage{
daysWithinSurface(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Tibble
}
\description{
Generate table showing the count and percentage of days per water year where water is within 10 cm and 50 cm of the ground surface (for years where there are at least 300 days of data)
}
