% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tables.R
\name{discreteVisitData}
\alias{discreteVisitData}
\title{Generate table showing depth to water measurements collected during site visits at a well}
\usage{
discreteVisitData(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Tibble
}
\description{
Generate table showing depth to water measurements collected during site visits at a well
}
