% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_formatting.R
\name{discreteVisitDataReactable}
\alias{discreteVisitDataReactable}
\title{Standardize formatting of discrete well visit data table returned by reactable package}
\usage{
discreteVisitDataReactable(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Reactable HTML object
}
\description{
Standardize formatting of discrete well visit data table returned by reactable package
}
