% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{loadWetlandWells}
\alias{loadWetlandWells}
\title{Load data into package environment}
\usage{
loadWetlandWells(
  data_path = c("database", "aquarius"),
  use_default_sql = TRUE,
  sql_drv = odbc::odbc()
)
}
\arguments{
\item{data_path}{Path or URL to the data. Accepted inputs:
\itemize{
\item \code{"aquarius"} (connect to the Aquarius database)
\item \code{"database"} (connect to the Access database)
}}

\item{use_default_sql}{Use default SQL database? Ignored if \code{data_path != "database"}.}

\item{sql_drv}{Driver to use to connect to database. Ignored if \code{data_path != "database"}.}
}
\value{
Invisibly return a list containing all data
}
\description{
Load data into package environment
}
