% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_formatting.R
\name{loggerBaroDeploymentsReactable}
\alias{loggerBaroDeploymentsReactable}
\title{Standardize formatting of baro logger deployments table returned by reactable package}
\usage{
loggerBaroDeploymentsReactable(site)
}
\arguments{
\item{site}{Three- or four-digit site code (e.g., KMMH, 315)}
}
\value{
Reactable HTML object
}
\description{
Standardize formatting of baro logger deployments table returned by reactable package
}
