% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tables.R
\name{loggerDeployments}
\alias{loggerDeployments}
\title{Generate table showing logger deployments at a well or baro site}
\usage{
loggerDeployments(well, site)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}

\item{site}{Three- or four-digit site code (e.g., KMMH, 315)}
}
\value{
Tibble
}
\description{
Generate table showing logger deployments at a well or baro site
}
