% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_formatting.R
\name{loggerWellDeploymentsReactable}
\alias{loggerWellDeploymentsReactable}
\title{Standardize formatting of well logger deployments table returned by reactable package}
\usage{
loggerWellDeploymentsReactable(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Reactable HTML object
}
\description{
Standardize formatting of well logger deployments table returned by reactable package
}
