% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_plots.R
\name{plotDaysWithinSurface}
\alias{plotDaysWithinSurface}
\title{Generate bar plot of percentage of days per water year where water is within 10 cm and 50 cm of the ground surface (for years where there are at least 300 days of data)}
\usage{
plotDaysWithinSurface(well)
}
\arguments{
\item{well}{Three-digit well code (e.g., 319)}
}
\value{
Ggplot object
}
\description{
Generate bar plot of percentage of days per water year where water is within 10 cm and 50 cm of the ground surface (for years where there are at least 300 days of data)
}
