% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{readAndFilterData}
\alias{readAndFilterData}
\title{Read wetland well data from all sources and filter by name of data frame}
\usage{
readAndFilterData(data.name)
}
\arguments{
\item{data.name}{Name of the analysis view containing the data. E.g., "TimeseriesWaterLevel", "Site". See details for full list of data name options.}
}
\value{
Tibble of filtered data
}
\description{
Read wetland well data from all sources and filter by name of data frame
}
\details{
\code{data.name} options are: Site, Visit, TimeseriesWaterLevel, TimeseriesRawPressure, TimeseriesBaroPressure, TimeseriesWaterTemperature
}
