% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_utils.R
\name{readAquarius}
\alias{readAquarius}
\title{Read continuous data from Aquarius}
\usage{
readAquarius(username = "aqreadonly", password = "aqreadonly")
}
\arguments{
\item{username}{Aquarius username}

\item{password}{Aquarius password}
}
\value{
List of tibbles
}
\description{
Read continuous data from Aquarius
}
