% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tables.R
\name{siteInformation}
\alias{siteInformation}
\title{Generate table showing site information for wells and baro sites}
\usage{
siteInformation(park, type)
}
\arguments{
\item{park}{DEPO, SEKI, YOSE}

\item{type}{Well, Baro}
}
\value{
Tibble
}
\description{
Generate table showing site information for wells and baro sites
}
