% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_formatting.R
\name{siteInformationBaroReactable}
\alias{siteInformationBaroReactable}
\title{Standardize formatting of baro site information table returned by reactable package}
\usage{
siteInformationBaroReactable(park)
}
\arguments{
\item{park}{Four-digit park code (e.g., YOSE)}
}
\value{
Reactable HTML object
}
\description{
Standardize formatting of baro site information table returned by reactable package
}
