% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_formatting.R
\name{siteInformationWellReactable}
\alias{siteInformationWellReactable}
\title{Standardize formatting of well site information table returned by reactable package}
\usage{
siteInformationWellReactable(park)
}
\arguments{
\item{park}{}
}
\value{
Reactable HTML object
}
\description{
Standardize formatting of well site information table returned by reactable package
}
